Crate gravatar

rust-gravatar is a small Rust library that generates Gravatar image URLs based on the official Gravatar specification.

Example

extern crate gravatar;
use gravatar::{Gravatar, Rating};

let mut g = Gravatar::new("email@example.com");
g.size = Some(150);
g.rating = Some(Rating::Pg);
assert_eq!(
    g.image_url(),
    "https://secure.gravatar.com/avatar/5658ffccee7f0ebfda2b226238b1eb6e\
     ?s=150&r=pg"
);

Structs

Gravatar
Representation of a single Gravatar image URL.

Enums

Default
The default image to display if the user's email does not have a Gravatar.

Rating
The maximum rating level for which Gravatar will show the user's image instead of the specified default.