Enum gravatar::Default

pub enum Default {
    Url(String),
    Http404,
    Blank,
    MysteryMan,
    Identicon,
    MonsterId,
    Wavatar,
    Retro,
}

The default image to display if the user's email does not have a Gravatar.

See http://en.gravatar.com/site/implement/images/#default-image.

Variants

Url

The URL of an image file to display as the default.

Http404

Instead of loading an image, the Gravatar URL will return an HTTP 404 (File Not Found) response if the email is not found.

Blank

A transparent PNG image.

MysteryMan

A simple, cartoon-style silhouetted outline of a person that does not vary by email hash.

Identicon

A geometric pattern based on the email hash.

MonsterId

A "monster" with different colors, faces, etc. that are generated by the email hash.

Wavatar

A face with different features and backgrounds, generated by the email hash.

Retro

An 8-bit arcade-style pixelated face that is generated by the email hash.